Use a context offset separate from the cursor.
It is frequently useful to retrieve a context for a range of
another context (e.g. with CONTEXT-SUBSEQ), and the range of
the resulting context shouldn't vary based on what has been
read. CONTEXT-DATA and CONTEXT-SUBSEQ will use the context's
start offset by default; to get the previous behavior, callers
can pass the current cursor as the start argument.