Use new stream abstraction for parser contexts.
Rather than assuming we're operating on a string, do all the parsing
from an abstract stream ("mini-streams"). This allows more complex
parsing behavior (for example parsing a stream that is being run
through a preprocessing stage), as well as allowing the use of unusual
data sources (like a list or array) and potentially non-character data
atoms. Performance over large inputs should also be better.

The previous code for constructing parser contexts from streams was
left intact, although it really won't perform well since it reads in
the entire stream regardless of the bounds used. The mini-stream type
that wraps a normal stream requires seekable streams, where the
existing code doesn't, so we can't safely convert it.

To support custom implementations of the stream abstraction, a new
method has been added to create a parser context out of a
mini-stream. This method notably doesn't support the :START and :END
bounds options, because those are typically present in the mini-stream
constructor.