Don't read more data than might be parsed.
When constructing a parser context from a stream, if an end bound was
specified, we don't need to read any data beyond that position,
because it's all EOF out there. Rather than reading the entire stream
contents up and then constructing a parser context for it, stop when
we've passed the end and use that portion of the data.

Note that if the specified range is small, but near the end of the
stream, we're still going to read in a large portion of data that will
never be parsed. We could use a similar trick to avoid storing leading
data we don't need, but then the cursor positions wouldn't match the
stream positions, and some callers may care about that.